// src/modul-2/async-project/interfaces/Location.ts
export interface Locations {
  info: Info
  results: Location[]
}

export interface Info {
  count: number
  pages: number
  next: string | null
  prev: string | null
}

export interface Location {
  id: number
  name: string
  type: string
  dimension: string
  residents: string[]
  url: string
  created: Date
}